XRDDEV-880 Upload anchor // PR fixes 3
* add tests
* move global conf download to GlobalConfService for easier mocking
